requirements.txt
Flask

numpy

Pillow

tensorflow

supabase-py

python-dotenv
gunicorn